fix leaking windows, move to AboutWindow
